requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
asyncer==0.0.14
attrs==25.4.0
backoff==2.2.1
bidict==0.23.1
certifi==2026.1.4
cffi==2.0.0
chainlit==2.9.6
charset-normalizer==3.4.4
chevron==0.14.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
cuid==0.4
dataclasses-json==0.6.7
deepeval==3.8.6
Deprecated==1.3.1
distro==1.9.0
docstring_parser==0.17.0
execnet==2.1.2
fastapi==0.129.0
ffmpeg-python==0.2.0
filelock==3.24.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
future==1.0.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-core==1.2.13
langchain-text-splitters==1.1.0
langsmith==0.7.3
Lazify==0.4.0
literalai==0.1.201
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.26.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
ollama==0.6.1
openai==2.24.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-agno==0.52.3
opentelemetry-instrumentation-alephalpha==0.52.3
opentelemetry-instrumentation-anthropic==0.52.3
opentelemetry-instrumentation-bedrock==0.52.3
opentelemetry-instrumentation-chromadb==0.52.3
opentelemetry-instrumentation-cohere==0.52.3
opentelemetry-instrumentation-crewai==0.52.3
opentelemetry-instrumentation-google-generativeai==0.52.3
opentelemetry-instrumentation-groq==0.52.3
opentelemetry-instrumentation-haystack==0.52.3
opentelemetry-instrumentation-lancedb==0.52.3
opentelemetry-instrumentation-langchain==0.52.3
opentelemetry-instrumentation-llamaindex==0.52.3
opentelemetry-instrumentation-logging==0.60b1
opentelemetry-instrumentation-marqo==0.52.3
opentelemetry-instrumentation-mcp==0.52.3
opentelemetry-instrumentation-milvus==0.52.3
opentelemetry-instrumentation-mistralai==0.52.3
opentelemetry-instrumentation-ollama==0.52.3
opentelemetry-instrumentation-openai==0.52.3
opentelemetry-instrumentation-openai-agents==0.52.3
opentelemetry-instrumentation-pinecone==0.52.3
opentelemetry-instrumentation-qdrant==0.52.3
opentelemetry-instrumentation-redis==0.60b1
opentelemetry-instrumentation-replicate==0.52.3
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-sagemaker==0.52.3
opentelemetry-instrumentation-sqlalchemy==0.60b1
opentelemetry-instrumentation-threading==0.60b1
opentelemetry-instrumentation-together==0.52.3
opentelemetry-instrumentation-transformers==0.52.3
opentelemetry-instrumentation-urllib3==0.60b1
opentelemetry-instrumentation-vertexai==0.52.3
opentelemetry-instrumentation-voyageai==0.52.3
opentelemetry-instrumentation-watsonx==0.52.3
opentelemetry-instrumentation-weaviate==0.52.3
opentelemetry-instrumentation-writer==0.52.3
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.60b1
orjson==3.11.7
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
portalocker==3.2.0
posthog==5.4.0
propcache==0.4.1
protobuf==6.33.5
psycopg==3.3.2
psycopg-binary==3.3.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
pyfiglet==1.0.4
Pygments==2.19.2
PyJWT==2.11.0
PyMuPDF==1.27.1
pypdf==6.7.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-repeat==0.9.4
pytest-rerunfailures==16.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-multipart==0.0.22
python-socketio==5.16.1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
sentry-sdk==2.53.0
setuptools==82.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.2.0
starlette==0.52.1
sympy==1.14.0
syncer==2.0.3
tabulate==0.9.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.10.0
tqdm==4.67.3
traceloop-sdk==0.52.3
transformers==5.2.0
triton==3.6.0
typer==0.23.2
typer-slim==0.23.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.41.0
voyageai==0.3.7
watchfiles==1.1.1
wheel==0.46.3
wrapt==1.17.3
wsproto==1.3.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
